Remove the Admin class
Remove the Admin class and move its methods to JobManager.
