+++ docs/reference.rst
@@ -62,13 +62,18 @@ Exceptions
     :members: ProcrastinateException, LoadFromPathError,
               ConnectorException, AlreadyEnqueued, AppNotOpen
 
-Administration
---------------
+Job statuses
+------------
 
-.. autoclass:: procrastinate.admin.Admin
-    :members: list_jobs, list_jobs_async, list_queues, list_queues_async,
-              list_tasks, list_tasks_async, set_job_status, set_job_status_async
+.. autoclass:: procrastinate.jobs.Status
+    :members:
 
+
+Accessing the jobs in the Database
+----------------------------------
+
+.. autoclass:: procrastinate.manager.JobManager
+    :members:
 
 Django
 ------
