+++ docs/spelling_wordlist.txt
@@ -1,4 +1,3 @@
-admin
 app
 async
 asynchronicity
